move rollout windows configuration to separate file
